insta-archiver/requirements.txt
instaloader
fpdf2
Pillow
